chore: 모듈 이름 변경, 오타 수정, 파일 분리
